Migrate calculator implementation to LangChain Tool subclass.
